requirements.txt
streamlit==1.28.1
openai==1.3.5
python-dotenv==1.0.0
sentence-transformers==2.3.0
chromadb==0.4.10
numpy==1.24.3
pandas==2.1.1
faiss-cpu==1.7.4
pydantic==1.10.12
requests==2.31.0
huggingface-hub==0.16.4
